Improve zsh.sh install script

+++ install/zsh.sh
@@ -1,10 +1,10 @@
 #!/usr/bin/env bash
 # ----------------------------------------------------------------------------#
-#                      _           _        _ _       _
-#                     (_)_ __  ___| |_ __ _| | |  ___| |__
-#                     | | '_ \/ __| __/ _` | | | / __| '_ \
-#                     | | | | \__ \ || (_| | | |_\__ \ | | |
-#                     |_|_| |_|___/\__\__,_|_|_(_)___/_| |_|
+#                                   _           _
+#                           _______| |__    ___| |__
+#                          |_  / __| '_ \  / __| '_ \
+#                           / /\__ \ | | |_\__ \ | | |
+#                          /___|___/_| |_(_)___/_| |_|
 #
 # Author: Serhat Teker <serhat.teker@gmail.com>
 # Source: https://github.com/SerhatTeker/dotfiles
@@ -24,15 +24,17 @@ ROOT="$( cd "$( dirname "${BASH_SOURCE[0]}" )/.." && pwd )"
 source "${ROOT}/install/common.sh"
 
 # Ask for the administrator password upfront
-sudo -v
+# sudo -v
 
-if [ -f "${XDG_CONFIG_HOME}/zsh" ] || [ ! -L "${XDG_CONFIG_HOME}/zsh" ];then
-    # or remove dir after informing in advance?
-    echo "${XDG_CONFIG_HOME}/zsh is not a symlink. Delete it manually."
-    exit 1
-else
-    unlink "${XDG_CONFIG_HOME}/zsh"
-    ln -sf "${DOTFILES}/.config/zsh" "${XDG_CONFIG_HOME}/zsh"
+if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+    if [ -f "${XDG_CONFIG_HOME}/zsh" ] || [ ! -L "${XDG_CONFIG_HOME}/zsh" ];then
+        # or remove dir after informing in advance?
+        echo "${XDG_CONFIG_HOME}/zsh is not a symlink. Delete it manually."
+        exit 1
+    else
+        unlink "${XDG_CONFIG_HOME}/zsh"
+        ln -sf "${DOTFILES}/.config/zsh" "${XDG_CONFIG_HOME}/zsh"
+    fi
 fi
 
 # Export main environment variables for ZSH
@@ -40,19 +42,28 @@ export ZMAIN=${XDG_CONFIG_HOME}/zsh
 export ZDOTDIR=${ZMAIN}
 export ZSH=${ZMAIN}/.oh-my-zsh
 
-
 default-shell() {
     # make zsh default shell
     sudo sh -c "echo $(which zsh) >> /etc/shells"
     chsh -s $(which zsh)
 }
 
-ioh-my-zsh() {
-    rm -rf ${DOTFILES}/.config/zsh/.oh-my-zsh
+# create soft links
+link-configs() {
+    ln -s ${DOTFILES}/.config/zsh/ ${XDG_CONFIG_HOME}
+    ln -sf ${SYSBAK}/zsh/.private.zsh ${ZMAIN}/.private.zsh
+    ln -sf ${PRIVATE}/zsh/.private.zsh ${ZMAIN}/.zsh_history
+}
 
-    wget https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh
-    ZSH="${ZMAIN}/.oh-my-zsh" sh install.sh
-    rm install.sh
+ioh-my-zsh() {
+    if [[ -d "${HOME}/dotfiles/.config/zsh/.oh-my-zsh" ]];then
+        rm -rf ${DOTFILES}/.config/zsh/.oh-my-zsh
+    fi
+    wget \
+        https://raw.githubusercontent.com/ohmyzsh/ohmyzsh/master/tools/install.sh \
+        -P /tmp
+    ZSH="${ZMAIN}/.oh-my-zsh" sh /tmp/install.sh
+    # ZSH="${ZMAIN}/.oh-my-zsh" sh /tmp/install.sh > /dev/null 2>&1
 }
 
 # TODO: replace/sed ZDOTDIR with XDG_CACHE_HOME in oh-my-zsh.sh
@@ -63,6 +74,8 @@ modify-oh-my-zsh() {
     echo ""
 }
 
+# TODO: Fix: not working together with ioh-my-zsh
+# Modify install.sh script
 # install custom plugins
 iplugins() {
     ZSH_CUSTOM="${ZSH}/custom"
@@ -84,32 +97,35 @@ iplugins() {
 
     # zsh-completions
     # https://github.com/zsh-users/zsh-completions
-    git clone https://github.com/zsh-users/zsh-completions
+    git clone https://github.com/zsh-users/zsh-completions \
         ${ZSH_CUSTOM:=~/.oh-my-zsh/custom}/plugins/zsh-completions
 }
 
 set-zdotdir() {
     # Set global ZDOTDIR
     # Hacky ugly way to fix tmux behavior
-    echo "export ZDOTDIR=\"\$HOME/.config/zsh\"" | \
-        sudo tee -a /etc/zsh/zshenv
-}
-
-# create soft links
-link-configs() {
-    ln -s ${DOTFILES}/.zshenv ${HOME}/.zshenv
-    ln -s ${DOTFILES}/.config/zsh/ ${XDG_CONFIG_HOME}
-    ln -sf ${SYSBAK}/zsh/.private.zsh ${ZMAIN}/.private.zsh
-    ln -sf ${PRIVATE}/zsh/.private.zsh ${ZMAIN}/.zsh_history
+    if [[ "$OSTYPE" == "linux-gnu"* ]]; then
+        echo "export ZDOTDIR=\"\$HOME/.config/zsh\"" | \
+            sudo tee -a /etc/zsh/zshenv
+    elif [[ "$OSTYPE" == "darwin"* ]]; then
+        # TODO: Implement
+        echo "export ZDOTDIR=\"\$HOME/.config/zsh\"" | \
+            sudo tee -a /etc/zshenv
+    else
+        echo "No install configuration for ${OSTYPE}"
+        exit 1
+    fi
 }
 
 main() {
+    # default-shell
+    set-zdotdir
     link-configs
     ioh-my-zsh
     iplugins
-    set-zdotdir
 }
 
-# main
+main
+# iplugins
 
 exit 0
